requirements.txt
ansicon==1.89.0
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Deprecated==1.2.14
dnspython==2.6.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Sitemap==0.4.0
Flask-WTF==1.2.1
greenlet==3.1.0
idna==3.8
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jinxed==1.3.0
jsonify==0.5
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-whois==0.9.4
redis==5.0.8
requests==2.32.3
rich==13.8.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2
